Make invalid movement check more lenient when walking up stairs/slabs
Instead of applying the default 0.125 value on the movement check, the
code first checks if there is a stair or slab nearby, if yes, the value
is increased to 0.19 (if he isn't sprinting) or 0.32 (if he is
sprinting).

I used 0.19 and 0.32 because these were the max value I got in a vanilla
MCW10 client when trying to walk on a stair.

After walking in a stair up and down for hours, I think the check is now
working better than before, with less rubberbanding issues.